+++ api/src/test/java/org/yapp/unit/domain/auth/application/oauth/google/GoogleOauthClientTest.java
@@ -1,4 +1,4 @@
-package org.yapp.domain.auth.application.oauth.google;
+package org.yapp.unit.domain.auth.application.oauth.google;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -16,6 +16,7 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.yapp.core.exception.ApplicationException;
 import org.yapp.core.exception.error.code.AuthErrorCode;
+import org.yapp.domain.auth.application.oauth.google.GoogleOauthClient;
 
 @ExtendWith(MockitoExtension.class)
 class GoogleOauthClientTest {
