+++ api/src/test/java/org/yapp/unit/domain/auth/application/profile/ProfileServiceTest.java
