+++ api/src/test/java/org/yapp/unit/domain/auth/application/term/TermUseCaseTest.java
